requirements.txt
appnope==0.1.3
backcall==0.2.0
#backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
confluent-kafka==1.9.2
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
idna==3.4
ipykernel==6.15.3
ipython==7.34.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kafka-python==1.4.7
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
oauthlib==3.2.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-twitter==3.5
pyzmq==24.0.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
